+++ lib/libftprintf/libft/srcs/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: fcorvaro <fcorvaro@student.42roma.it>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/17 18:45:00 by fcorvaro          #+#    #+#             */
-/*   Updated: 2023/11/15 14:08:39 by fcorvaro         ###   ########.fr       */
+/*   Updated: 2024/06/30 13:06:11 by fcorvaro         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,9 +21,9 @@
  *
  * @return int Returns the length of the string.
  */
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
